backend/requirements.txt
flask
flask-cors
requests
pillow
numpy
tensorflow-cpu==2.19.0
